+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Added `HorizontalLayout` manager (#228)
 - Added `POINT_X, POINT_Y` and `SIZE_WIDTH, SIZE_HEIGHT` `usize` constants for `Points` and `Size` types.
+- Added `PaddingConstraint` struct.
+- Added `CONFIG_PADDING` option to `WidgetConfig` class.
 
 ## 0.4.13